+++ NEWS.md
@@ -19,7 +19,7 @@ Note: Newly deprecated arguments will continue working during a transition perio
 Enhancements:
   * 'tz' is now checked against the output of OlsonNames() to ensure a valid timezone is entered.
   * messages and warnings are now printed using message() and warning()
-
+  * Stop if any of the input files has duplicated column names
 
 ## actel 0.1.1